dev-requirements.txt
coverage
minio==2.2.2
mock
requests

# PyTest for running the tests.
pytest==3.0.5
pytest-django==3.1.2
pytest-cov==2.4.0
pytest-pythonpath
